style: 캘린더 배경 디자인 수정
