+++ src/pages/calendar/CalendarPage.js
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import { Link, useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
 
 import Calendar from "../../components/calendar/Calendar";
 import Schedule from "../../components/calendar/Schedule";
@@ -14,50 +14,7 @@ import { BiUserCircle } from "react-icons/bi";
 
 import color from "./../../assets/color.png";
 
-// Background_circle
-const Circle_frame = styled.div`
-  width: 25px;
-  height: 25px;
-  border-radius: 50%;
-  background-color: #8165df;
-  opacity: 0;
-  @keyframes wave {
-    0% {
-      scale: 100%;
-      opacity: 0.8;
-    }
-    40% {
-      opacity: 0.3;
-    }
-    100% {
-      scale: 2000%;
-      opacity: 0;
-    }
-  }
-`;
-const L_Circle_frame = styled(Circle_frame)`
-  position: absolute;
-  bottom: -25px;
-  left: -25px;
-`;
 
-const R_Circle_frame = styled(Circle_frame)`
-  position: absolute;
-  top: -25px;
-  right: -25px;
-`;
-const Circle1 = styled(L_Circle_frame)`
-  animation: wave 4s infinite linear;
-`;
-const Circle2 = styled(L_Circle_frame)`
-  animation: wave 4s infinite 2s linear;
-`;
-const Circle3 = styled(R_Circle_frame)`
-  animation: wave 4s infinite linear;
-`;
-const Circle4 = styled(R_Circle_frame)`
-  animation: wave 4s infinite 2s linear;
-`;
 
 const Test = styled.div`
   width: 100vw;
@@ -125,13 +82,6 @@ const Last = styled.div`
 const CalendarPage = () => {
   return (
     <Test>
-      <div className="circle_frame" style={{ overflow: "hidden" }}>
-        <Circle1 />
-        <Circle2 />
-        <Circle3 />
-        <Circle4 />
-      </div>
-
       <MainDiv className="MainDiv">
         <BackColor src={color} style={{ opacity: 0.2 }} />
         <Header>
